validate auth + specs
